boost::posix_time::seconds constructor requires integer argument

+++ ublox_gps/include/ublox_gps/node.h
@@ -477,7 +477,7 @@ typedef boost::shared_ptr<ComponentInterface> ComponentPtr;
 class UbloxNode : public virtual ComponentInterface {
  public:
   //! How long to wait during I/O reset [s]
-  constexpr static double kResetWait = 10.0;
+  constexpr static int kResetWait = 10;
   //! how often (in seconds) to call poll messages
   constexpr static double kPollDuration = 1.0;
   // Constants used for diagnostic frequency updater
